Add module volcano plot
Since I don't have the DESeq2 output from the upstream module I used the
"old_script_example_output/Morbus Crohn_Control_IHWsigGenes.xlsx" file
to run the script. The output directory was randomly chosen and can be
easily changed.

When using this data no labels are being displayed on the volcano
plot as none of the "genes_of_interest" match the symbols in the DESeq2
output. This can be easily checked by adding an available gene such as
"MUC6".

Also, it would be easy to add a title to the volcano plot provided that
one would come up with an inspirational name ...

The inspiration for passing arguments to an R script from command lines:

https://www.r-bloggers.com/2015/09/passing-arguments-to-an-r-script-from-command-lines/